+++ bt/app/ui/picker_app.h
@@ -14,6 +14,16 @@ namespace bt::ui {
             ImVec2 max;
         };
 
+        struct action_menu_item {
+            std::string id;
+            std::string icon;
+            std::string tooltip;
+            float x{0};
+            float y{0};
+            float angle{0};
+            float angle_final{0};
+        };
+
 
         picker_app(const std::string& url);
         ~picker_app();
@@ -30,6 +40,7 @@ namespace bt::ui {
         const float ProfileSquareSize = 40.0f;
         const float ProfileSquarePadding = 5.0f;
         const float InactiveProfileSquarePadding = 7.0f;
+        const float ActionSquareSize = 20.0f;
         const float WindowMinWidth = BrowserSquareSize * 6;
         const float WindowHeight = BrowserSquareSize + ProfileSquareSize + 80;
         float wnd_width;
@@ -51,11 +62,18 @@ namespace bt::ui {
         std::vector<connection_box> profiles_cb; // active profile coordinates
         float browser_bar_left_pad{0};
 
+        bool action_menu_hovered{false};
+        std::vector<action_menu_item> action_menu_items{
+            action_menu_item{"copy", ICON_MD_CONTENT_COPY, "Copy to clipboard & close"},
+            action_menu_item{"email", ICON_MD_EMAIL, "Email link"}
+        };
+
         grey::widgets::window wnd_main;
 
         std::shared_ptr<bt::browser_instance> decision;
 
         bool run_frame();
+        void render_action_menu(float x, float y);
         void render_browser_bar();
         void render_profile_bar();
         void render_connection_box();
@@ -65,5 +83,7 @@ namespace bt::ui {
          * @param decision 
          */
         void make_decision(std::shared_ptr<bt::browser_instance> decision);
+
+        void menu_item_clicked(const std::string& id);
     };
 }